% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crime_counts.R
\name{crime_counts}
\alias{crime_counts}
\title{extract crime counts}
\usage{
crime_counts(x, y, yr)
}
\arguments{
\item{x}{tibble of cps data}

\item{y}{tibble of cpd data}
}
\description{
extract crime counts
}
