% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fast.as.Date}
\alias{fast.as.Date}
\title{fast Date function}
\usage{
fast.as.Date(x, format = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{x}{character string}

\item{format}{date/time parsing format}
}
\description{
fast Date function
}
