% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fast.as.IDate}
\alias{fast.as.IDate}
\title{fast IDate function}
\usage{
fast.as.IDate(x, format = "\%Y-\%m-\%d", ...)
}
\arguments{
\item{x}{character string}

\item{format}{date/time parsing format}
}
\description{
fast IDate function
}
