% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fast.as.ITime}
\alias{fast.as.ITime}
\title{fast ITime function}
\usage{
fast.as.ITime(x, format = "\%H:\%M:\%S", ...)
}
\arguments{
\item{x}{character string}

\item{format}{date/time parsing format}
}
\description{
fast ITime function
}
