% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpd-data.R
\name{get_cpd_crime}
\alias{get_cpd_crime}
\title{get chicago public school data}
\usage{
get_cpd_crime(data_dir = file.path(get_main_dir(), "RawData/Chicago_Crime/"),
  force = FALSE, replace = FALSE)
}
\arguments{
\item{data_dir}{path to data folder.}

\item{force}{force a full import.}

\item{replace}{replace existing package data with new import.}
}
\description{
get chicago public school data
}
