% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpd-data.R
\name{get_cpd_crime_data}
\alias{get_cpd_crime_data}
\title{get chicago public school data}
\usage{
get_cpd_crime_data(data_dir = file.path(get_main_dir(),
  "RawData/Chicago_Crime/"), import = FALSE)
}
\arguments{
\item{data_dir}{path to data folder.}

\item{force}{force a full import.}
}
\description{
get chicago public school data
}
