% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-data.R
\name{get_cps_address}
\alias{get_cps_address}
\title{get chicago public school data}
\usage{
get_cps_address(data_dir = file.path(get_main_dir(), "Data"), force = FALSE)
}
\arguments{
\item{data_dir}{path to data folder.}

\item{force}{force a full import.}
}
\description{
get chicago public school data
}
