% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-data.R
\name{get_cps_data}
\alias{get_cps_data}
\title{get chicago public school data}
\usage{
get_cps_data(raw_data_dir = "~/Dropbox/ra-work/spped/Data/", force = FALSE)
}
\description{
get chicago public school data
}
