% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-data.R
\name{get_cps_personnel}
\alias{get_cps_personnel}
\title{get chicago public school data. REQUIRES DATA CLEANED BY SCRIPT `bin/clean-cps-personnel.sh`.}
\usage{
get_cps_personnel(raw_data_dir = file.path(get_main_dir(),
  "RawData/CPS_Personnel/cooked"), force = FALSE)
}
\arguments{
\item{raw_data_dir}{path to raw data folder.}

\item{force}{force a full import.}
}
\description{
get chicago public school data. REQUIRES DATA CLEANED BY SCRIPT `bin/clean-cps-personnel.sh`.
}
