% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-data.R
\name{get_cps_security_personnel}
\alias{get_cps_security_personnel}
\title{get chicago public school data. REQUIRES DATA CLEANED BY SCRIPT `bin/clean-cps-personnel.sh`.}
\usage{
get_cps_security_personnel(data_dir = file.path(get_main_dir(),
  "RawData/CPS_Personnel/cooked"), force = FALSE, replace = FALSE)
}
\arguments{
\item{data_dir}{path to data directory}

\item{force}{force a full import.}

\item{replace}{replace existing package data with new import.}
}
\description{
get chicago public school data. REQUIRES DATA CLEANED BY SCRIPT `bin/clean-cps-personnel.sh`.
}
