% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crime_counts.R
\name{get_crime_counts}
\alias{get_crime_counts}
\title{count crimes found within a .25 and .5 miles of a school}
\usage{
get_crime_counts(force = FALSE, replace = FALSE, export = TRUE)
}
\arguments{
\item{force}{whether to recreate the entire cpd_crime csv file}

\item{replace}{whether to replace the entire cpd_crime csv file}

\item{export}{whether to export files}
}
\description{
count crimes found within a .25 and .5 miles of a school
}
