% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crime_counts.R
\name{get_within_miles}
\alias{get_within_miles}
\title{determine counts within .25 and .5 miles by crime type}
\usage{
get_within_miles(l, y, yr)
}
\arguments{
\item{l}{vector with (lon, lat) coords}

\item{y}{tibble of cpd data}
}
\description{
determine counts within .25 and .5 miles by crime type
}
