% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_main_dir}
\alias{has_main_dir}
\title{Make sure sep = "|", encoding = "UTF-8") main_dir string is not empty}
\usage{
has_main_dir()
}
\value{
logical TRUE if main_dir is not empty
}
\description{
Make sure sep = "|", encoding = "UTF-8") main_dir string is not empty
}
