% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-data.R
\name{get_cps_address_data}
\alias{get_cps_address_data}
\title{get chicago public school data}
\usage{
get_cps_address_data(raw_data_dir = "~/Dropbox/ra-work/spped/Data/",
  force = FALSE)
}
\arguments{
\item{raw_data_dir}{path to raw data folder.}

\item{force}{force a full import.#' @export}
}
\description{
get chicago public school data
}
