% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps-data.R
\name{get_cps_personnel}
\alias{get_cps_personnel}
\title{get chicago public school data. REQUIRES DATA CLEANED BY SCRIPT `bin/clean-cps-personnel.sh`.}
\usage{

  get_cps_personnel(data_dir = "~/Dropbox/ra-work/spped/RawData/CPS_Personnel/cooked",
  force = FALSE)
}
\arguments{
\item{data_dir}{path to data folder.}

\item{force}{force a full import.}
}
\description{
get chicago public school data. REQUIRES DATA CLEANED BY SCRIPT `bin/clean-cps-personnel.sh`.
}
